test and fix einstein momentum on gpu vs cpu. long double not supported in cuda.

+++ SimulationEngineLib/shared/service/momentum/einstein/EinsteinMomentumService.cuh
@@ -37,8 +37,8 @@ protected:
     ) override;
 
 private:
-    const long double speedLight = PhysicalConstants::SPEED_OF_LIGHT;
-	const long double speedLightSquared = speedLight * speedLight;
+    const double speedLight = PhysicalConstants::SPEED_OF_LIGHT;
+	const double speedLightSquared = speedLight * speedLight;
 
     float getGamma(
         Vector3D<float> velocity
